Move AssertUtil to an external jar
