+++ assertions-enabled/pom.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <version>2.23ea1-SNAPSHOT</version>
+    <groupId>net.openhft</groupId>
+    <artifactId>assertions-enabled</artifactId>
+    <packaging>jar</packaging>
+
+    <name>OpenHFT/Chronicle-Assertions-Enabled</name>
+    <description>Chronicle-Assertions-Enabled</description>
+
+</project>
