+++ pom.xml
@@ -55,6 +55,14 @@
     </dependencyManagement>
 
     <dependencies>
+
+        <!--TODO: Use profile to select type-->
+        <dependency>
+            <groupId>net.openhft</groupId>
+            <artifactId>assertions-enabled</artifactId>
+            <version>2.23ea1-SNAPSHOT</version>
+        </dependency>
+
         <dependency>
             <groupId>net.openhft</groupId>
             <artifactId>affinity</artifactId>
