+++ src/main/java/net/openhft/chronicle/core/UnsafeMemory.java
@@ -32,7 +32,7 @@ import java.nio.ByteOrder;
 import java.util.concurrent.atomic.AtomicLong;
 
 import static java.util.Objects.nonNull;
-import static net.openhft.chronicle.core.util.AssertUtil.SKIP_ASSERTIONS;
+import static net.openhft.chronicle.assertions.AssertUtil.SKIP_ASSERTIONS;
 import static net.openhft.chronicle.core.util.Ints.assertIfEnabled;
 import static net.openhft.chronicle.core.util.Longs.assertIfEnabled;
 import static net.openhft.chronicle.core.util.ObjectUtils.requireNonNull;
