+++ src/main/java/net/openhft/chronicle/core/util/Ints.java
@@ -1,5 +1,6 @@
 package net.openhft.chronicle.core.util;
 
+import net.openhft.chronicle.assertions.AssertUtil;
 import net.openhft.chronicle.core.internal.invariant.ints.IntBiCondition;
 import net.openhft.chronicle.core.internal.invariant.ints.IntCondition;
 import net.openhft.chronicle.core.internal.invariant.ints.IntTriCondition;
